requirements.txt
datasets>=2.14.0
python-dotenv>=1.0.0
pandas>=2.0.0
jupyter>=1.0.0
notebook>=7.0.0
regex>=2025.11.0
tqdm>=4.65.0
